requirements.txt
fastapi
uvicorn[standard]
gspread
google-auth
jinja2
python-multipart
pandas
python-dotenv
